Fix asset property ordering…again

+++ digiroad2-oracle/src/main/scala/fi/liikennevirasto/digiroad2/asset/oracle/OracleSpatialAssetDao.scala
@@ -62,7 +62,7 @@ object OracleSpatialAssetDao {
       val row = v(0)
       AssetWithProperties(id = row.id, assetTypeId = row.assetTypeId,
         lon = row.lon, lat = row.lat, roadLinkId = row.roadLinkId,
-        propertyData = AssetPropertyConfiguration.assetRowToCommonProperties(row) ++ assetRowToProperty(v).sortBy(_.propertyId),
+        propertyData = AssetPropertyConfiguration.assetRowToCommonProperties(row) ++ assetRowToProperty(v).sortBy(_.propertyId.toLong),
         bearing = row.bearing, municipalityNumber = Option(row.municipalityNumber),
         validityPeriod = validityPeriod(row.validFrom, row.validTo),
         imageIds = v.map(row => getImageId(row.image)).toSeq.filter(_ != null),
